Update authenticate.py
if the header is None then returning None will cause the function to stop right there and the code will not be read that is below the return None. If we have a raw_token already from the cookie then it will not be even used. So don't return anything in here.